make navigator widget factory extensible
