+++ packages/navigator/src/browser/index.ts
@@ -16,4 +16,5 @@
 
 export * from './navigator-model';
 export * from './navigator-widget';
+export * from './navigator-widget-factory';
 export * from './navigator-decorator-service';
