+++ packages/navigator/src/browser/navigator-frontend-module.ts
@@ -19,14 +19,14 @@ import '../../src/browser/style/index.css';
 import { ContainerModule } from 'inversify';
 import {
     KeybindingContext, bindViewContribution,
-    FrontendApplicationContribution, ViewContainer,
+    FrontendApplicationContribution,
     ApplicationShellLayoutMigration
 } from '@theia/core/lib/browser';
-import { FileNavigatorWidget, FILE_NAVIGATOR_ID, EXPLORER_VIEW_CONTAINER_ID, EXPLORER_VIEW_CONTAINER_TITLE_OPTIONS } from './navigator-widget';
+import { FileNavigatorWidget, FILE_NAVIGATOR_ID } from './navigator-widget';
 import { NavigatorActiveContext } from './navigator-keybinding-context';
 import { FileNavigatorContribution } from './navigator-contribution';
 import { createFileNavigatorWidget } from './navigator-container';
-import { WidgetFactory, WidgetManager } from '@theia/core/lib/browser/widget-manager';
+import { WidgetFactory } from '@theia/core/lib/browser/widget-manager';
 import { bindFileNavigatorPreferences } from './navigator-preferences';
 import { FileNavigatorFilter } from './navigator-filter';
 import { NavigatorContextKeyService } from './navigator-context-key-service';
@@ -35,6 +35,7 @@ import { NavigatorDiff } from './navigator-diff';
 import { NavigatorLayoutVersion3Migration } from './navigator-layout-migrations';
 import { NavigatorTabBarDecorator } from './navigator-tab-bar-decorator';
 import { TabBarDecorator } from '@theia/core/lib/browser/shell/tab-bar-decorator';
+import { NavigatorWidgetFactory } from './navigator-widget-factory';
 
 export default new ContainerModule(bind => {
     bindFileNavigatorPreferences(bind);
@@ -55,22 +56,8 @@ export default new ContainerModule(bind => {
         id: FILE_NAVIGATOR_ID,
         createWidget: () => container.get(FileNavigatorWidget)
     })).inSingletonScope();
-    bind(WidgetFactory).toDynamicValue(({ container }) => ({
-        id: EXPLORER_VIEW_CONTAINER_ID,
-        createWidget: async () => {
-            const viewContainer = container.get<ViewContainer.Factory>(ViewContainer.Factory)({
-                id: EXPLORER_VIEW_CONTAINER_ID,
-                progressLocationId: 'explorer'
-            });
-            viewContainer.setTitleOptions(EXPLORER_VIEW_CONTAINER_TITLE_OPTIONS);
-            const widget = await container.get(WidgetManager).getOrCreateWidget(FILE_NAVIGATOR_ID);
-            viewContainer.addWidget(widget, {
-                canHide: false,
-                initiallyCollapsed: false
-            });
-            return viewContainer;
-        }
-    }));
+    bind(NavigatorWidgetFactory).toSelf().inSingletonScope();
+    bind(WidgetFactory).toService(NavigatorWidgetFactory);
     bind(ApplicationShellLayoutMigration).to(NavigatorLayoutVersion3Migration).inSingletonScope();
 
     bind(NavigatorDiff).toSelf().inSingletonScope();
